requirements.txt
uvicorn==0.30.1
fastapi==0.110.3
python-dotenv==1.0.0
crewai==0.28.8
langchain==0.1.20
langchain-core==0.1.52
langchain-aws==0.1.3
sqlalchemy==2.0.31
tiktoken==0.7.0
pydantic==2.8.2
async-timeout==4.0.3
psycopg2-binary==2.9.9
anthropic